+++ flask_dance/contrib/dropbox.py
@@ -1,0 +1,82 @@
+from __future__ import unicode_literals
+
+from flask import request, url_for
+from flask_dance.consumer import OAuth2ConsumerBlueprint
+from functools import partial
+from flask.globals import LocalProxy, _lookup_app_object
+try:
+    from flask import _app_ctx_stack as stack
+except ImportError:
+    from flask import _request_ctx_stack as stack
+
+
+__maintainer__ = "David Baumgold <david@davidbaumgold.com>"
+
+
+def make_dropbox_blueprint(
+        app_key=None, app_secret=None, scope=None,
+        force_reapprove=False, disable_signup=False,
+        redirect_url=None,
+        redirect_to=None, login_url=None, authorized_url=None,
+        session_class=None, backend=None):
+    """
+    Make a blueprint for authenticating with Dropbox using OAuth 2. This requires
+    a client ID and client secret from Dropbox. You should either pass them to
+    this constructor, or make sure that your Flask application config defines
+    them, using the variables DROPBOX_OAUTH_APP_KEY and DROPBOX_OAUTH_APP_SECRET.
+
+    Args:
+        app_key (str): The client ID for your application on Dropbox.
+        app_secret (str): The client secret for your application on Dropbox
+        scope (str, optional): comma-separated list of scopes for the OAuth token
+        redirect_url (str): the URL to redirect to after the authentication
+            dance is complete
+        redirect_to (str): if ``redirect_url`` is not defined, the name of the
+            view to redirect to after the authentication dance is complete.
+            The actual URL will be determined by :func:`flask.url_for`
+        login_url (str, optional): the URL path for the ``login`` view.
+            Defaults to ``/github``
+        authorized_url (str, optional): the URL path for the ``authorized`` view.
+            Defaults to ``/github/authorized``.
+        session_class (class, optional): The class to use for creating a
+            Requests session. Defaults to
+            :class:`~flask_dance.consumer.requests.OAuth2Session`.
+        backend: A storage backend class, or an instance of a storage
+                backend class, to use for this blueprint. Defaults to
+                :class:`~flask_dance.consumer.backend.session.SessionBackend`.
+
+    :rtype: :class:`~flask_dance.consumer.OAuth2ConsumerBlueprint`
+    :returns: A :ref:`blueprint <flask:blueprints>` to attach to your Flask app.
+    """
+    authorization_url_params = {}
+    if force_reapprove:
+        authorization_url_params["force_reapprove"] = "true"
+    if disable_signup:
+        authorization_url_params["disable_signup"] = "true"
+
+    dropbox_bp = OAuth2ConsumerBlueprint("dropbox", __name__,
+        client_id=app_key,
+        client_secret=app_secret,
+        scope=scope,
+        base_url="https://api.dropbox.com/1/",
+        authorization_url="https://www.dropbox.com/1/oauth2/authorize",
+        token_url="https://api.dropbox.com/1/oauth2/token",
+        redirect_url=redirect_url,
+        redirect_to=redirect_to,
+        login_url=login_url,
+        authorized_url=authorized_url,
+        authorization_url_params=authorization_url_params,
+        session_class=session_class,
+        backend=backend,
+    )
+    dropbox_bp.from_config["client_id"] = "DROPBOX_OAUTH_APP_KEY"
+    dropbox_bp.from_config["client_secret"] = "DROPBOX_OAUTH_APP_SECRET"
+
+    @dropbox_bp.before_app_request
+    def set_applocal_session():
+        ctx = stack.top
+        ctx.dropbox_oauth = dropbox_bp.session
+
+    return dropbox_bp
+
+dropbox = LocalProxy(partial(_lookup_app_object, "dropbox_oauth"))