+++ tests/consumer/test_oauth2.py
@@ -119,6 +119,7 @@ def test_authorized_url():
         assert len(responses.calls) == 1
         request_data = dict(parse_qsl(responses.calls[0].request.body))
         assert request_data["client_id"] == "client_id"
+        assert request_data["redirect_uri"] == "https://a.b.c/login/test-service/authorized"
         # check that we stored the access token and secret in the session
         assert (
             flask.session["test-service_oauth_token"] ==